+++ attributes/default.rb
@@ -14,3 +14,12 @@
 # limitations under the License.
 #
 
+# Disabled by default, sample ports local port 2222 to remote localhost port 22
+default['autossh-init']['ssh_lports'] = ''
+#default['autossh-init']['ssh_lports'] = '-Llocalhost:2222:localhost:22'
+
+# Disabled by default, sample ports remote host port 2222 to localhost 22
+default['autossh-init']['ssh_rports'] = ''
+#default['autossh-init']['ssh_lports'] = '-Rlocalhost:2022:localhost:22'
+
+